+++ R/esc_beta.R
@@ -10,12 +10,14 @@
 #' @param es.type Type of effect size that should be returned.
 #'        \describe{
 #'          \item{\code{"d"}}{returns standardized mean difference effect size \code{d}}
+#'          \item{\code{"f"}}{returns effect size Cohen's \code{f}}
 #'          \item{\code{"g"}}{returns adjusted standardized mean difference effect size Hedges' \code{g}}
 #'          \item{\code{"or"}}{returns effect size as odds ratio}
 #'          \item{\code{"cox.or"}}{returns effect size as Cox-odds ratio (see \code{\link{esc_d2or}} for details)}
 #'          \item{\code{"logit"}}{returns effect size as log odds}
 #'          \item{\code{"cox.log"}}{returns effect size as Cox-log odds (see \code{\link{esc_d2logit}} for details)}
 #'          \item{\code{"r"}}{returns correlation effect size \code{r}}
+#'          \item{\code{"eta"}}{returns effect size eta squared}
 #'        }
 #' @param study Optional string with the study name. Using \code{\link{combine_esc}} or
 #'        \code{as.data.frame} on \code{esc}-objects will add this as column
@@ -39,7 +41,7 @@
 #'
 #' @export
 esc_beta <- function(beta, sdy, grp1n, grp2n,
-                     es.type = c("d", "g", "or", "logit", "r", "cox.or", "cox.log"),
+                     es.type = c("d", "g", "or", "logit", "r", "f", "eta", "cox.or", "cox.log"),
                      study = NULL) {
   # match  arguments
   es.type <- match.arg(es.type)
@@ -53,6 +55,13 @@ esc_beta <- function(beta, sdy, grp1n, grp2n,
   v <- esc.vd(es, grp1n, grp2n)
 
   # return effect size
-  return(esc_generic(es = es, v = v, es.type = es.type, grp1n = grp1n, grp2n = grp2n,
-                     info = "standardized regression coefficient", study = study))
+  esc_generic(
+    es = es,
+    v = v,
+    es.type = es.type,
+    grp1n = grp1n,
+    grp2n = grp2n,
+    info = "standardized regression coefficient",
+    study = study
+  )
 }